backend/requirements.txt
fastapi
uvicorn
python-jose

python-jose
passlib[bcrypt]
psycopg2-binary

